# Add clamp(val, min, max) function to pygame.math