--- a/buildconfig/stubs/pygame/math.pyi
+++ b/buildconfig/stubs/pygame/math.pyi
@@ -18,6 +18,8 @@ if sys.version_info >= (3, 9):
 else:
     from typing import Collection
 
+def clamp(value: float, min: float, max: float) -> float: ...
+
 _VectorTypeVar = TypeVar("_VectorTypeVar", Vector2, Vector3)
 
 # VectorElementwiseProxy is a generic, it can be an elementwiseproxy object for