--- a/src_c/math.c
+++ b/src_c/math.c
@@ -4048,28 +4048,20 @@ vector_elementwise(pgVector *vec, PyObject *_null)
 }
 
 static PyObject *
-math_clamp(PyObject *self, PyObject *arg)
+math_clamp(PyObject *self, PyObject *const *args, Py_ssize_t nargs)
 {
-    if (PyTuple_Size(arg) != 3)
+    if (nargs != 3)
         return RAISE(PyExc_ValueError, "clamp requires 3 float arguments");
 
-    // Parse val argument
-    PyObject *val = PyTuple_GET_ITEM(arg, 0);
-    double cval = PyFloat_AsDouble(val);
-
-    // Parse minval argument
-    PyObject *minval = PyTuple_GET_ITEM(arg, 1);
-    double cminval = PyFloat_AsDouble(minval);
-
-    // Parse maxval argument
-    PyObject *maxval = PyTuple_GET_ITEM(arg, 2);
-    double cmaxval = PyFloat_AsDouble(maxval);
+    double value = PyFloat_AsDouble(args[0]);
+    double min = PyFloat_AsDouble(args[1]);
+    double max = PyFloat_AsDouble(args[2]);
 
     if (PyErr_Occurred()) {
         return RAISE(PyExc_TypeError, "clamp requires 3 float arguments");
     }
 
-    double result = fmin(cmaxval, fmax(cminval, cval));
+    double result = fmin(max, fmax(min, value));
 
     return PyFloat_FromDouble(result);
 }
@@ -4101,7 +4093,7 @@ math_disable_swizzling(pgVector *self, PyObject *_null)
 }
 
 static PyMethodDef _math_methods[] = {
-    {"clamp", math_clamp, METH_VARARGS, DOC_PYGAMEMATHCLAMP},
+    {"clamp", math_clamp, METH_FASTCALL, DOC_PYGAMEMATHCLAMP},
     {"enable_swizzling", (PyCFunction)math_enable_swizzling, METH_NOARGS,
      "Deprecated, will be removed in a future version"},
     {"disable_swizzling", (PyCFunction)math_disable_swizzling, METH_NOARGS,