--- a/src_c/math.c
+++ b/src_c/math.c
@@ -4050,18 +4050,22 @@ vector_elementwise(pgVector *vec, PyObject *_null)
 static PyObject *
 math_clamp(PyObject *self, PyObject *const *args, Py_ssize_t nargs)
 {
+    const char *err_msg = "clamp requires 3 float arguments";
+
     if (nargs != 3)
-        return RAISE(PyExc_ValueError, "clamp requires 3 float arguments");
+        return RAISE(PyExc_ValueError, err_msg);
 
     double value = PyFloat_AsDouble(args[0]);
     if (PyErr_Occurred())
-        return RAISE(PyExc_TypeError, "clamp requires 3 float arguments");
+        return RAISE(PyExc_TypeError, err_msg);
+
     double min = PyFloat_AsDouble(args[1]);
     if (PyErr_Occurred())
-        return RAISE(PyExc_TypeError, "clamp requires 3 float arguments");
+        return RAISE(PyExc_TypeError, err_msg);
+
     double max = PyFloat_AsDouble(args[2]);
     if (PyErr_Occurred())
-        return RAISE(PyExc_TypeError, "clamp requires 3 float arguments");
+        return RAISE(PyExc_TypeError, err_msg);
 
     double result = fmin(max, fmax(min, value));