--- a/src_c/math.c
+++ b/src_c/math.c
@@ -109,6 +109,11 @@ typedef struct {
 } vector_elementwiseproxy;
 
 /* further forward declarations */
+/* math functions */
+static PyObject *
+math_clamp(PyObject *self, PyObject *const *args, Py_ssize_t nargs);
+PG_DECLARE_FASTCALL_FUNC(math_clamp, PyObject);
+
 /* generic helper functions */
 static int
 RealNumber_Check(PyObject *obj);
@@ -4070,14 +4075,7 @@ math_clamp(PyObject *self, PyObject *const *args, Py_ssize_t nargs)
     return PyFloat_FromDouble(result);
 }
 
-#if PY_VERSION_HEX < 0x03070000
-static PyObject *
-math_clamp_wrapper(PyObject *self, PyObject *args)
-{
-    return math_clamp(self, (PyObject *const *)PySequence_Fast_ITEMS(args),
-                      PyTuple_GET_SIZE(args));
-}
-#endif
+PG_WRAP_FASTCALL_FUNC(math_clamp, PyObject);
 
 static PyObject *
 math_enable_swizzling(pgVector *self, PyObject *_null)
@@ -4106,12 +4104,8 @@ math_disable_swizzling(pgVector *self, PyObject *_null)
 }
 
 static PyMethodDef _math_methods[] = {
-#if PY_VERSION_HEX < 0x03070000
-    {"clamp", (PyCFunction)math_clamp_wrapper, METH_VARARGS,
+    {"clamp", (PyCFunction)PG_FASTCALL_NAME(math_clamp), PG_FASTCALL,
      DOC_PYGAMEMATHCLAMP},
-#else
-    {"clamp", (PyCFunction)math_clamp, METH_FASTCALL, DOC_PYGAMEMATHCLAMP},
-#endif
     {"enable_swizzling", (PyCFunction)math_enable_swizzling, METH_NOARGS,
      "Deprecated, will be removed in a future version"},
     {"disable_swizzling", (PyCFunction)math_disable_swizzling, METH_NOARGS,