--- a/test/math_test.py
+++ b/test/math_test.py
@@ -14,17 +14,77 @@ class MathModuleTest(unittest.TestCase):
 
     def test_clamp(self):
         """Test clamp function."""
+
+        # Int tests
+        # Test going above max
         result = pygame.math.clamp(10, 1, 5)
-        self.assertAlmostEqual(result, 5)
-
+        self.assertEqual(result, 5)
+        # Test going below min
         result = pygame.math.clamp(-10, 1, 5)
-        self.assertAlmostEqual(result, 1)
-
+        self.assertEqual(result, 1)
+        # Test equal to max
         result = pygame.math.clamp(5, 1, 5)
-        self.assertAlmostEqual(result, 5)
-
+        self.assertEqual(result, 5)
+        # Test equal to min
         result = pygame.math.clamp(1, 1, 5)
-        self.assertAlmostEqual(result, 1)
+        self.assertEqual(result, 1)
+        # Test between min and max
+        result = pygame.math.clamp(3, 1, 5)
+        self.assertEqual(result, 3)
+
+        # Float tests
+        # Test going above max
+        result = pygame.math.clamp(10.0, 1.12, 5.0)
+        self.assertAlmostEqual(result, 5.0)
+        # Test going below min
+        result = pygame.math.clamp(-10.0, 1.12, 5.0)
+        self.assertAlmostEqual(result, 1.12)
+        # Test equal to max
+        result = pygame.math.clamp(5.0, 1.12, 5.0)
+        self.assertAlmostEqual(result, 5.0)
+        # Test equal to min
+        result = pygame.math.clamp(1.12, 1.12, 5.0)
+        self.assertAlmostEqual(result, 1.12)
+        # Test between min and max
+        result = pygame.math.clamp(2.5, 1.12, 5.0)
+        self.assertAlmostEqual(result, 2.5)
+
+        # Point < and > are calculated based on the magnitude.
+        class Point:
+            def __init__(self, x, y):
+                self.x = x
+                self.y = y
+
+            def length(self):
+                return math.sqrt(self.x**2 + self.y**2)
+
+            def __lt__(self, other):
+                return self.length() < other.length()
+
+            def __gt__(self, other):
+                return self.length() > other.length()
+
+            def __eq__(self, other):
+                if isinstance(other, Point):
+                    return (self.x == other.x) and (self.y == other.y)
+                return False
+
+        # Point tests
+        # Test going above max
+        result = pygame.math.clamp(Point(100, 100), Point(2, 2), Point(40, 3))
+        self.assertTrue(result == Point(40, 3))
+        # Test going below min
+        result = pygame.math.clamp(Point(0, 0), Point(2, 2), Point(40, 3))
+        self.assertTrue(result == Point(2, 2))
+        # Test equal to max
+        result = pygame.math.clamp(Point(40, 3), Point(2, 2), Point(40, 3))
+        self.assertTrue(result == Point(40, 3))
+        # Test equal to min
+        result = pygame.math.clamp(Point(2, 2), Point(2, 2), Point(40, 3))
+        self.assertTrue(result == Point(2, 2))
+        # Test between min and max
+        result = pygame.math.clamp(Point(10, 10), Point(2, 2), Point(40, 3))
+        self.assertTrue(result == Point(10, 10))
 
 
 class Vector2TypeTest(unittest.TestCase):