I know that it's simple to implement your own clamp function, but clamp seems to be used in games quite a bit. If you're on the fence for whether or not this should be included with pygame, rather than just implemented in python, here are a few points:

- Most other frameworks implement their own clamp function (MonoGame, Unity, Godot, even the C++ standard library)
- There aren't any (that I know of) popular python libraries that have simple functions like clamp.
- Obviously, since pygame.math.clamp is implemented in C, it should be faster than a python implementation. I timed it, and pygame.math.clamp is around 75% faster than an equivalent python implementation, while (I believe) providing all the same behaviour.

The function can take arguments of any type, which may seem strange, but it is equivalent to how it would work if it was implemented in python, _and_ it's much faster than enforcing a specific type, which requires converting the python type to C type, which slows it down a bit.

I'm new to contributing, so I'm sorry if I did something wrong. I couldn't find the test stub generator, so I manually wrote it.