@itzpr3d4t0r 2022-07-20
you are missing the math.pyi function declaration + this could be a nice function do make with FASTCALL: info #3324.