@JamesC01 2022-07-20
> 

Is That as simple as changing from METH_VARARGS to METH_FASTCALL? Also, how do I generate the pyi?