@JamesC01 2022-07-20
> > Is that as simple as changing from METH_VARARGS to METH_FASTCALL? Also, how do I generate the pyi?
> 
> No you have to do a couple of things.
> 
> 1. Yes you have to change `METH_VARARGS` to `METH_FASTCALL`
> 2. You have to change the function signature to `static PyObject *math_clamp(PyObject *self, PyObject *const *args, Py_ssize_t nargs)` so now you have `args` which will be a C array of `PyObject*` that are your 3 values. `nargs` represents the number of arguments passed to the function so you use that to validate the correct arguments number.
> 3. Now instead of using `PyObject *val = PyTuple_GET_ITEM(arg, 0);` for example, you write `PyObject *val = args[0];` etc.
> 
> You don't generate a pyi. you have to write it yourself. find the math.pyi file and add your function. if you want to generate docs(like the ones in the .h file) what i do is just write the docs in the .rst file and then run this command: `python setup.py docs --fullgeneration`

Thanks, I'll get to it! What about generating test stubs? I couldn't find the file mentioned in the README.