@JamesC01 2022-07-20
> I mean i just add the test to the correct test file. In this case shouldn't it be math_test.py?

Yeah, I added it, but in test/README.md it says to use test/utils/gen_stubs.py to generate test stubs, but the file doesn't exist, so I had to write the test myself. It seems to work, though.