@JamesC01 2022-07-20
> Hello! First of all thanks for taking the time and effort for getting started with contributing to pygame.
> 
> A clamp function was already added as a vector method in #2990
> 
> You probably missed it because it's not on the pygame docs as there has been no pygame release since the time the clamp function was added

Isn't that for clamping vectors? The clamp I added is for clamping floats.