@itzpr3d4t0r 2022-07-20
Oh damn i forgot to tell you that FASTCALL is only supported from python 3.7 onwards, so you'll have to do two more things.
Add this under your function definition:
```C
#if PY_VERSION_HEX < 0x03070000
static PyObject *
math_clamp_wrapper(PyObject *self, PyObject *args) {
    return math_clamp(self, (PyObject *const *)PySequence_Fast_ITEMS(args), PyTuple_GET_SIZE(args));
}
#endif
```