@itzpr3d4t0r 2022-07-23
Now that #3325 has been merged it is recommended that you use that way to implement a FASTCALL function. Sorry to ask more work from you, but thiss will simplify the code, making it more readable and easier to delete the 3.6 part when we drop compat somewhere in the future.