@itzpr3d4t0r 2022-07-25
Actually i found out that you can make it even faster, i'll give you the code i used to test and some results:

### Proposed change
my proposed change/reorganisation:
static PyObject *
```C
math_clamp(PyObject *self, PyObject *const *args, Py_ssize_t nargs)
{
    double value, min, max;
    if (nargs != 3)
        return RAISE(PyExc_ValueError, "clamp requires 3 float arguments");

    value = PyFloat_Check(args[0]) ? PyFloat_AS_DOUBLE(args[0])
                                   : (double)PyLong_AsLong(args[0]);
    if (PyErr_Occurred())
        return NULL;
    min = PyFloat_Check(args[1]) ? PyFloat_AS_DOUBLE(args[1])
                                 : (double)PyLong_AsLong(args[1]);
    if (PyErr_Occurred())
        return NULL;
    max = PyFloat_Check(args[2]) ? PyFloat_AS_DOUBLE(args[2])
                                 : (double)PyLong_AsLong(args[2]);
    if (PyErr_Occurred())
        return NULL;

    return PyFloat_FromDouble(MIN(max, MAX(min, value)));
}
```

___
The results:
**CURRENT**
```
====||MATH.CLAMP||====
-) 1: 0.68651996
-) 2: 0.7033555
-) 3: 0.69078892
______________
-) Mean time: 0.69355479

====||PYTHON  CLAMP||====
-) 1: 2.17853738
-) 2: 2.17602866
-) 3: 2.10800866
______________
-) Mean time: 2.15419157
```
**WITH MY PROPOSED CHANGE**
```
====||MATH.CLAMP||====
-) 1: 0.55778852
-) 2: 0.56320468
-) 3: 0.56477658
______________
-) Mean time: 0.56192326

====||PYTHON CLAMP||====
-) 1: 2.18498714
-) 2: 2.22100118
-) 3: 2.13322032
______________
-) Mean time: 2.17973621
```
The code:

```Python
import timeit
from statistics import fmean

from pygame.math import clamp


def myclamp(x, minv, maxv):
    return min(maxv, max(minv, x))


NUMBER = 10_000_000
tottime = 0
GLOB = {"clamp": clamp, "myclamp": myclamp}

def test(test_name: str, func: str) -> float:
    value = fmean(timeit.repeat(func, globals=GLOB, number=NUMBER))
    print("-) " + test_name + f": {round(value, 8)}")
    return value


def test_group(
        group_name: str,
        sequence: list[str],
        tests_names: tuple[str, ...] = None,
        include_tot: bool = False,
        include_mean: bool = True
) -> float:
    print("\n====||" + group_name.upper() + "||====")
    
    test_data = [test(tests_names[i] if tests_names else str(i + 1), test_obj)
                 for
                 i, test_obj in
                 enumerate(sequence)]
    
    print("______________")
    if include_tot:
        print(f"-) Total time: {round(sum(test_data), 8)}")
    
    if include_mean:
        print(f"-) Mean time: {round(fmean(test_data), 8)}")
    
    return fmean(test_data)


test_group(
    "math.clamp",
    ["clamp(1, 15, 17)",
     "clamp(22, 15, 17)",
     "clamp(8, 4, 17)",
     ])

test_group(
    "python clamp",
    ["myclamp(1, 15, 17)",
     "myclamp(22, 15, 17)",
     "myclamp(8, 4, 17)"]
)
```