@MyreMylar 2022-07-31
> > LGTM. Gave it a good round of testing locally and ran the unit tests provided.
> > Approved! 👍 💯 🥂 🎉
> 
> Thanks! Did you take a look at the unit tests? I'm very knew to this, so I'm not sure if they're very good. They seem to work. I just check different cases, making sure they give the correct value, but since I'm new, I could be missing something.

I did take a look, they seemed fine. Could maybe have tested the invalid class/object case, but otherwise good 👍